fix : host_age bigint로 변경시도
Former-commit-id: 874e6ca9a20e19dd7e6790b469a3701d07632feb

+++ exec/docker/initial-mysql/init.sql
@@ -53,7 +53,7 @@ create table home (
       guardian_name varchar(255),
       guardian_phone varchar(255),
       host_account_no varchar(255),
-      host_age integer,
+      host_age bigint,
       host_bank varchar(255),
       host_gender varchar(255),
       host_name varchar(255),
